Add new hook to check whether Debian security uploads have been authorized
We used to have the Debian security upload queues disabled by default
to avoid accidental uploads by users that had not read, or had not be
authorized to perform such uploads.

Add a new hook that gives the necessary information and aks whether
the user has been authorized. This way we can uncomment the queues by
default, which means users will get updates to the config.