Avoid explicit usage of the topic variable
The topic variable makes sense when it it is required as part of a
function interface, or when it is only implicitly used. If we have
to use it explicitly then there is something wrong going on, and a
clearly named variable is always a better option.

Addresses: BuiltinFunctions::ProhibitUselessTopic
Warned-by: perlcritic