Add support for multiple hooks per category
This will make possible to specify multiple hooks, instead of chaining
them from within the hooks themselves, which does not scale, when there
are several of them, and we need to reorder or specify a subset per
host.

For backwards compatibility we will still allow single hooks per
category while emitting a warning.