Fix announce address concatenation handling
When accumulating the «To» addresses for an announcement, we were
concatenating into an undef string, separating entries with spaces, and
ending up always with a leading spurious space. Which would not be valid
in case of multiple addresses.

Instead switch it to an array, and then format when we are preparing the
output, by concatenating addresses with «, ».

Warned-by: perl