hooks: Add new Debian transition check hook
This new hook for Debian target hosts, checks the transition information
and will prompt whenever the «.changes» to upload is involved in one or
more ongoing transitions.

This hook ignores «.changes» files that are not targeting the Debian
unstable distribution, or that are not sourceful (such as uploads from
build daemons.

Ref: #1060367