hooks: Rewrite openpgp-check in perl for Dpkg::OpenPGP multi-backend support
This switches from the hardcoded GnuPG handling to automatically and
transparently supporting in addition both SOP and Sequoia-PGP if available.
It will use the vendor keyrings to verify the upload instead of whatever
might be in the GnuPG keystore.

This is actually a reintroduction of this change, which had to be
reverted, as there was previously no easy way to specify the keyrings to
use to check the signatures against.