debian: Add Recommends on debian-keyring | ubuntu-keyring
These are now needed as part of the OpenPGP hook, and due to the default
hosts configured in dupload.conf.

In theory this should be using vendor-neutral dependency, because the
list of keyrings to use comes indirectly from the Dpkg::Vendor::<vendor>
module, but given that we explicitly have dupload configuration for
these to vendors, it does make sense to be explicit about the keyring
packages we expected to have installed.

Closes: #1099556